Print view mostly working
I can't figure out how to get it to not cut off in the middle of an element...From what I've found, it's basically not possible to ensure that in Chrome >_>

Also, I started working on getting searching working by clicking on a name from the list--should have committed that seperately, but I...forgot I hadn't committed the print change. Woops.