Oh my word. Printing is better.
At least in Chrome. I can't speak to other browsers yet, I'm afraid. My frustration of page-break-inside: avoid; not working was due to the fact that I was using flex as the display. Of course, nowhere that I had seen this had mentioned that it was a problem until I looked it up specifically ("page-break-inside webkit flex"). And then it was still broken until I figured out the drawing pane needed to be position: absolute; top: 0px; to get it up to the top and get the page breaks in the right place in one fell swoop.

Only took me...an hour this go round.